use visiblitychange event
